MTP

The MTP.jl package is for working with SBS microplates as tidy DataFrames.

MTP.LETTERS

DataFrames.DataFrame

MTP.merge_info

MTP.plate

MTP.platesetup

MTP.platesetupfile

MTP.printplate

MTP.wells

MTP.LETTERS — Constant

MTP.LETTERS is just the vector ["A", "B", ..., "Z"] as in R

DataFrames.DataFrame — Method

DataFrame(s::XLSX.Worksheet)
Get a work-sheet as DataFrame

MTP.merge_info — Method

merge_info(df::DataFrame)
Add Q, well96, well384, row96, col96, row384, col384 to a plate DataFrame

MTP.plate — Function

MTP.plate(platename, geometry = 96)
MTP.plate([plate1, plate2], geometry = 96)
create one or more plates with the given names and geometry.

MTP.platesetup — Method

platesetup(df::DataFrame)
Get at plate-setup from a DataFrame

MTP.platesetupfile — Method

platesetupfile(file::String)
Get a platesetup for an .xlsx file.
Reads all sheets.
Returns DataFrame of filename, sheetname, platename, geometry, well, well_content
Platename is written in the topleft corner. well_content is in the wells.
Only first plate definition on a sheet is read (for now)

MTP.printplate — Function

printplate(df::DataFrame, dispcol= :well)
Show data in plate format. The dispcol selects the color to show in the wells.
Eg MTP.printplate(su3, :Q)

MTP.wells — Method

wells(geometry)
wells(rows, cols)

Generate wells based on geometry or rows and colums (A01, A02, ....)